Move flask to seperate file and create Procfile
