+++ Procfile
@@ -1,0 +1,2 @@
+web: gunicorn flask_app:app
+worker: python3 core.py
